# docs/api-swagger/swagger.yaml
...
info:
  contact: {}
paths:
...
  /golish/api/artifacts:
    get:
      description: Get a paginated list of artifacts with optional filtering and existence
        checks
      parameters:
      - description: Filter by workspace name
        in: query
        name: workspace
        type: string
      - description: Search in artifact name/path
        in: query
        name: search
        type: string
      - description: Filter by HTTP status code (also accepts statusCode)
        in: query
        name: status_code
        type: integer
      - default: false
        description: Annotate results with path_exists and path_is_dir
        in: query
        name: verify_exist
        type: boolean
      - default: 0
        description: Number of records to skip
        in: query
        name: offset
        type: integer
      - default: 20
        description: Maximum number of records to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of artifacts with pagination
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to fetch artifacts
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List artifacts
      tags:
      - Artifacts
  /golish/api/artifacts/{workspace_name}:
    get:
      description: Download a single file under the given workspace by relative artifact
        path
      parameters:
      - description: Workspace name
        in: path
        name: workspace_name
        required: true
        type: string
      - description: Relative path to artifact under workspace
        in: query
        name: artifact_path
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Artifact file
          schema:
            type: file
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Artifact not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to download artifact
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Download workspace artifact
      tags:
      - Artifacts
...
  /golish/api/assets/diff:
    get:
      description: Compare assets between two time points to find added, removed,
        and changed assets
      parameters:
      - description: Workspace name
        in: query
        name: workspace
        required: true
        type: string
      - description: Start time (RFC3339 format or Unix timestamp)
        in: query
        name: from
        required: true
        type: string
      - description: 'End time (default: now)'
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Asset diff result
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid parameters
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to get asset diff
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get asset diff
      tags:
      - Assets
  /golish/api/assets/diffs:
    get:
      description: Get a paginated list of stored asset diff snapshots
      parameters:
      - description: Filter by workspace name
        in: query
        name: workspace
        type: string
      - default: 0
        description: Number of records to skip
        in: query
        name: offset
        type: integer
      - default: 20
        description: Maximum number of records to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of asset diff snapshots with pagination
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to fetch asset diff snapshots
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List asset diff snapshots
      tags:
      - Assets
...
  /golish/api/vulnerabilities/diff:
    get:
      description: Compare vulnerabilities between two time points to find added,
        removed, and changed vulnerabilities
      parameters:
      - description: Workspace name
        in: query
        name: workspace
        required: true
        type: string
      - description: Start time (RFC3339 format or Unix timestamp)
        in: query
        name: from
        required: true
        type: string
      - description: 'End time (default: now)'
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Vulnerability diff result
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid parameters
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to get vulnerability diff
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get vulnerability diff
      tags:
      - Vulnerabilities
  /golish/api/vulnerabilities/diffs:
    get:
      description: Get a paginated list of stored vulnerability diff snapshots
      parameters:
      - description: Filter by workspace name
        in: query
        name: workspace
        type: string
      - default: 0
        description: Number of records to skip
        in: query
        name: offset
        type: integer
      - default: 20
        description: Maximum number of records to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of vulnerability diff snapshots with pagination
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to fetch vulnerability diff snapshots
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List vulnerability diff snapshots
      tags:
      - Vulnerabilities
...
swagger: "2.0"